added atom mapping to forcite, discover, dftb and dmol
