+++ MDANSE_GUI/Src/MDANSE_GUI/Tabs/Visualisers/Action.py
@@ -39,6 +39,7 @@ widget_lookup = {  # these all come from MDANSE_GUI.InputWidgets
     "InputFileConfigurator": InputFileWidget,
     "MDFileConfigurator": InputFileWidget,
     "FieldFileConfigurator": InputFileWidget,
+    "XTDFileConfigurator": InputFileWidget,
     "XYZFileConfigurator": InputFileWidget,
     "AseInputFileConfigurator": AseInputFileWidget,
     "RunningModeConfigurator": RunningModeWidget,
